Выносит обёртку игрового экрана в отдельный компонент GameScreen
