+++ src/components/app/app.jsx
@@ -2,14 +2,12 @@ import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import {BrowserRouter, Route, Switch} from 'react-router-dom';
 
+import {GameType} from '@/const';
+
 import {WelcomeScreen} from '@/components/welcome-screen/welcome-screen';
+import {GameScreen} from '@/components/game-screen/game-screen';
 import {QuestionArtist} from '@/components/question-artist/question-artist';
 import {QuestionGenre} from '@/components/question-genre/question-genre';
-
-const GameType = {
-  ARTIST: `artist`,
-  GENRE: `genre`
-};
 
 class App extends PureComponent {
   constructor(props) {
@@ -40,26 +38,34 @@ class App extends PureComponent {
       switch (question.type) {
         case GameType.ARTIST:
           return (
-            <QuestionArtist
-              question={questions[1]}
-              onAnswer={() => {
-                this.setState((prevState) => ({
-                  step: prevState.step + 1
-                }));
-              }}
-            />
+            <GameScreen
+              type={question.type}
+            >
+              <QuestionArtist
+                question={questions[1]}
+                onAnswer={() => {
+                  this.setState((prevState) => ({
+                    step: prevState.step + 1
+                  }));
+                }}
+              />
+            </GameScreen>
           );
 
         case GameType.GENRE:
           return (
-            <QuestionGenre
-              question={questions[0]}
-              onAnswer={() => {
-                this.setState((prevState) => ({
-                  step: prevState.step + 1
-                }));
-              }}
-            />
+            <GameScreen
+              type={question.type}
+            >
+              <QuestionGenre
+                question={questions[0]}
+                onAnswer={() => {
+                  this.setState((prevState) => ({
+                    step: prevState.step + 1
+                  }));
+                }}
+              />
+            </GameScreen>
           );
       }
     }
