+++ src/components/question-genre/question-genre.jsx
@@ -16,69 +16,45 @@ class QuestionGenre extends PureComponent {
     const {answers: userAnswers} = this.state;
 
     return (
-      <section className="game game--genre">
-        <header className="game__header">
-          <a className="game__back" href="#">
-            <span className="visually-hidden">Сыграть ещё раз</span>
-            <img className="game__logo" src="img/melody-logo-ginger.png" alt="Угадай мелодию"/>
-          </a>
-
-          <svg xmlns="http://www.w3.org/2000/svg" className="timer" viewBox="0 0 780 780">
-            <circle className="timer__line" cx="390" cy="390" r="370"
-              style={{
-                filter: `url(#blur)`,
-                transform: `rotate(-90deg) scaleY(-1)`,
-                transformOrigin: `center`
-              }}/>
-          </svg>
-
-          <div className="game__mistakes">
-            <div className="wrong"></div>
-            <div className="wrong"></div>
-            <div className="wrong"></div>
-          </div>
-        </header>
-
-        <section className="game__screen">
-          <h2 className="game__title">Выберите {question.genre} треки</h2>
-          <form
-            className="game__tracks"
-            onSubmit={(evt) => {
-              evt.preventDefault();
-              onAnswer(question, this.state.answers);
-            }}
-          >
-            {answers.map((answer, index) => {
-              return (
-                <div key={answer.src} className="track">
-                  <button className="track__button track__button--play" type="button"></button>
-                  <div className="track__status">
-                    <audio></audio>
-                  </div>
-                  <div className="game__answer">
-                    <input
-                      className="game__input visually-hidden"
-                      type="checkbox"
-                      name="answer"
-                      value={`answer-${index + 1}`}
-                      id={`answer-${index + 1}`}
-                      onChange={(evt) => {
-                        const value = evt.target.checked;
-                        this.setState({
-                          answers: [...userAnswers.slice(0, index), value, ...userAnswers.slice(index + 1)]
-                        });
-                      }}
-                      checked={userAnswers[index]}
-                    />
-                    <label className="game__check" htmlFor={`answer-${index + 1}`}>Отметить</label>
-                  </div>
+      <section className="game__screen">
+        <h2 className="game__title">Выберите {question.genre} треки</h2>
+        <form
+          className="game__tracks"
+          onSubmit={(evt) => {
+            evt.preventDefault();
+            onAnswer(question, this.state.answers);
+          }}
+        >
+          {answers.map((answer, index) => {
+            return (
+              <div key={answer.src} className="track">
+                <button className="track__button track__button--play" type="button"></button>
+                <div className="track__status">
+                  <audio></audio>
                 </div>
-              );
-            })}
+                <div className="game__answer">
+                  <input
+                    className="game__input visually-hidden"
+                    type="checkbox"
+                    name="answer"
+                    value={`answer-${index + 1}`}
+                    id={`answer-${index + 1}`}
+                    onChange={(evt) => {
+                      const value = evt.target.checked;
+                      this.setState({
+                        answers: [...userAnswers.slice(0, index), value, ...userAnswers.slice(index + 1)]
+                      });
+                    }}
+                    checked={userAnswers[index]}
+                  />
+                  <label className="game__check" htmlFor={`answer-${index + 1}`}>Отметить</label>
+                </div>
+              </div>
+            );
+          })}
 
-            <button className="game__submit button" type="submit">Ответить</button>
-          </form>
-        </section>
+          <button className="game__submit button" type="submit">Ответить</button>
+        </form>
       </section>
     );
   }
