+++ src/const.js
@@ -1,0 +1,6 @@
+const GameType = {
+  ARTIST: `artist`,
+  GENRE: `genre`
+};
+
+export {GameType};
